requirements.txt
openai
langchain
chromadb
streamlit
beautifulsoup4
python-dotenv
